+++ src/main/java/com/atlauncher/data/MojangAccount.java
@@ -143,6 +143,10 @@ public class MojangAccount extends AbstractAccount {
     }
 
     @Override
+    public void changeSkinPreCheck() {
+    }
+
+    @Override
     public String getSkinUrl() {
         StringBuilder response;
         try {